requirements.txt
instapy==0.6.16
Pillow==9.3.0
requests==2.22.0
tweepy==4.12.1
